Add checks on return value of getTLVarray() and getI2CMRsps()
Macro CHECK_IF_ERROR_AND_ACCUMULATE is now used to check the return value
of getTLVarray(). If an error is foud the function returns with status APDU_ERROR
else offset le is updated.

Macro CHECK_IF_ERROR is now used to check the return value of getI2CMRsps().
If an error is foud the function returns with status APDU_ERROR.

Signed-off-by: Michael Grand <michael.grand.mg@gmail.com>