Fix memory leak on I2C driver object
In i2c.cpp an I2C object is created when axI2CInit() function is called.
The corresponding memory is not freed when connection with the SE050 is closed.
To fix this memory leak, a new function axI2CClose() is added and called by
phPalEse_i2c_close(). se050_disconnect() is modified to call only phNxpEse_close()
as phNxpEse_deInit() and phNxpEse_close() do the same thing which would result in a double free.

Signed-off-by: Michael Grand <michael.grand.mg@gmail.com>